UDP firmware update support for USRP2P.
The hooks are in there for USRP2, but without CPLD changes it won't support it.

Added an app host/utils/usrp2p_fw_update.py to write to USRP2P over the wire.

Lots of TODOs in that file. Caveat -- fw_common.h, bootloader_utils.h, and the .py app MUST ALL AGREE!