uhd: third iteration of the reference vector
Hope we get it right this time.
Reference vector now has its own pointer for the pointer/size 1 constuction case.
In this case, the memory is initialized to the value of its own pointer.
The previous two iterations were functionally wrong because it takes two pointers and size to accomplish this.