uhd: implemented boost barriers on all code that creates threads
The barrier ensures that the thread must spawn before the caller exits.

Some of the code already used a mutex to accomplish this,
however cygwin chokes when a mutex is locked twice by the same thread.
Mutex implementations were replaced with the barrier implementation.
Also the barrier implementation is far cleaner.