usrp2: allow devices to be discovered when usrp2_iface::make throws
If the discovery throws an exception, we can still yeild a discovered device.
The error (which is a problem) will re-appear at make/factory time.

The side effect is that a device w/ error will be discovered w/ empty serial and names.
We protect against this by supporing older firmwares where i2c proto is compatible.