usrp2: added the concept of device locking to usrp2/nseries devices
Each iface instance can lock itself to a particular device.
When the device is locked, it cannot be discovered through find.

Locking works by having a lock thread periodically update
a register in the firmware that says when the device was last locked.
The find routine can determine if a device is locked by checking
the difference between the current time and this register.