usrp2: added REF_MIMO and PPS_MIMO for USRP2 clock config
Basically, one can tell a MIMO master to lock to MIMO ref and/or time.
This is an alternative to the mimo_mode=slave/master device args.

When using this alternative, one should always set the mimo_mode to maser.
When in master mode, any clock config settings can be used.
When in slave mode, the clock config setting are forced to MIMO_REF/PPS