usrp1: fixed swapped sign on rx cordic
We used to swap I and Q to share converter functions with the newer products.
Because of this, the sign on the cordic also had to be swapped.

Now that USRP1 has its own converter routines and I and Q are correct,
so should be the sign on the cordic. This was a bug introduced in master.

TX should not be an issue, because there is no cordic
and is adjusted digitally by the codec, after any potential swapping.