uhd: branch-less round for time spec convert to ticks
This fixes valgrind warnings with branching on unit'd values,
and possible exceptions thrown on bad unit'd values.

As it turns out, sometimes tick_rate was passed unit'd,
however, the value calculated was not of consequence.

Also, iround was overkill because we dont need to handle
negative tick rates or frac seconds, or bounds checking.