vita: moved clear register to overlap with nchan register
This fixes the bug where setting the format clears the vita RX.
This is only an issue when the noclear option is set by UHD,
because the format register is always so, so it always clears.
Note: noclear is there to support the backwards compat API (pre streamer).

Now, numchans and clear overlap. This is ok because
in the host code, clear and numchans are always used together.

All timing meets on N2xx and USRP2.