xcvr2450: fix to disable LO offset for TX side
XCVR2450 has a common LO for RX and TX.

The use LO offset should be the same for both sides.
When different, a tune for one side will override the other side,
because the DSP will not know the compensation factor after the other side is tuned.

Setting both use_lo_offset to false is how this was pre 3.4 release.