utils: uhd_usrp_probe gets argument --init-only
This will run the tool without the output (unless --tree is provided).
Using --init-only, uhd_usrp_probe can be used to simply test a device
in more detail than doing a uhd_find_device, but the output of the
device initalization is not scrolled by because of the device info
output.