Crimson: Updated handle to reflect server.
The API on the Crimson server running on the device was heavily
modified. Thus changed had to be updated to the UHd driver to mate with
the new API. Tested to be working, and added additional error checks.

Crimson does not store frequency ranges on the device, nor eeprom
properties. We see fit that those properties should be stored within the
host side. Majority of the features of Crimson are hacked to fit within
the standards of the UHD driver. However, Per Vices is working to
support ALL the end features that Ettus devices also supports. It is a work
in progress.