FLOWCONTROL: Now only reads the fifo level if we are waiting for the timeslot to send data. Changed how timeslot is updated
There is basically 2 conditions in which the send command is called.
1) Faster than the sample rate requires in which case we need to wait
2) As fast as we need to send (or near it).
For the second condition, we do not have time to wait for a fifo read in which case we send data as fast as possible.
Otherwise if we need to wait anyway, we should use this time to read from fifo.
The timeslot used to be updated with the sample rate period upon udp transaction. Now it updates it as if it instantaniously
send the udp packet percicely when it needed to.