If we are cross compiling, do not test for build machine distro.
UHD would use the result of the REDHAT test to decide if UHD installed
in lib64. So if you did a build for a 64 bit machine on a FEDORA box with
OE, it decided to install uhd in /usr/lib64. This is not always the right thing
to do.

Eventually, we will need to deal with OE builds that use lib64, but we can figure
that out when we hit the problem.

Signed-off-by: Philip Balister <philip@opensdr.com>