utils: refactor uhd_images_downloader
Changing structure of the files.ettus.com images directory, and
updating uhd_images_downloader correspondingly. The images downloader
now downloads a number of smaller archives of images, instead of one
large archive containing all of our images. Furthermore, the images
downloader maintains an inventory of currently installed images, so
that images are not redownloaded unnecessarily.

When called with no arguments, behavior should not change.

CMake variables are used to populate a number of fields in the images
downloader, including the contents of the manifest file.

Reviewed-by: Martin Braun <martin.braun@ettus.com>
Reviewed-by: Ashish Chaudhari <ashish.chaudhari@ettus.com>