rpclib: fix use of "boost::asio::strand" -> "boost::asio::io_service::strand"
The former was marked as deprecated numerous Boost versions ago and
finally was actually removed & replaced in 1.66.0 with a new one with a
template API. The version in rpclib need to be updated, and Boost docs
say to use the latter. Moving to this usage takes care of this issue.

Reviewed-by: Martin Braun <martin.braun@ettus.com>