usrp2: Re-add ability to modulate in the DAC
This partially reverts b29e80cc. That commit fixed a bug with numerical
overflows, but also removed the ability to tune using the DAC,
effectively reducing the DSP tuning range.

This commit allows to tune within +/- 200 MHz using a combination of
both the DAC and the DSP tuning in the FPGA.

Reviewed-by: Derek Kozek <derek.kozel@ettus.com>