python: Adding phase alignment example
This tool uses the Python API to determine if USRP devices are
receiving a phase aligned signal. It is designed to be a stand-in
replacement for usrp_phasealignment.py, although the capabilities and
invocation are not identical.

The return value of this tool will indicate whether the signal is
sufficiently phase aligned, and allows integration into scripts or
other test frameworks. This tool also includes options to save data to
file or plot it using PyLab.

Currently only supports 2 channel RX phase alignment.