utils: uhd_images_downloader: Add download limit
This download limit prevents users from accidentally downloading really
large files. The default limit is 100 MiB. When a file is selected for
download that exceeds this size, the user is prompted if she really
wants to download that file.

By specifying --yes (or -y), this can can be bypassed. Example:

    $ uhd_images_downloader --yes -t sdimg

Will match all files that match 'sdimg', which are typically larger than
100 MiB, without interaction.