utils:rpc: set rpc timeout during rpc call
The current implementation of the UHD RPC client has a timeout that is
being accessed non-atomically. Many calls follow the pattern:
1. set_timeout(value)
2. request_rpc()
3. set_timeout(default)
which is not atomic. Other concurrent calls on the same rpc client may
change the timeout value; leads to unexpected behavior

These new set of function will, instead, handle
setting and re-setting the timeout atomically in the RPC request.