cpack: Fix RPM generation
UHDPackage.cmake produced new directory entries ("%dir" directives)
for /usr/local and other folders... This produced conflicts with the
filesystem package on Fedora 28 (at least).

Added set of CPACK_RPM_EXCLUDE_FROM_AUTO_FILELIST_ADDITION to the cmake
rpm package instructions.  This fixes the problem and produces a working
rpm for Fedora 28 x86_64.

The redhat post_install.in file was referring to the /usr/lib tree,
which is now /usr/lib64, so the script failed to install the
uhd-usrp.rules. That's been fixed too.