ad9361_device: Make time pauses check device status
At several points in this class, it waits for an arbitrary period of time, then checks whether the device is ready, and throws an exception if it isn't.  At other points, it uses a loop with pauses and a max index to check whether the device is ready and throw errors at timeout.  I rewrote these sections to instead be a loop checking a register with a specific timeout.  This means that the wait time is as short as possible.  The explicit timeout is more clear, and more importantly improving the loop while calibrating tx quadrature reduces retuning times of the B210 from 20-40 ms to 3-7 ms.