RFNoC: Remove incorrect flush
Removing the flush in configure_flow_control_out().  The flush is done
incorrectly because it just disables flow control and allows packets
to dump onto the crossbar, which could lock it up.  It is also
incorrect to flush when connecting blocks.  A connect should just
configure the SID and flow control and let any existing data flow to
the newly connected block.  Flushing of a block should only be done
during creation or destruction of the block.

Signed-off-by: Michael West <michael.west@ettus.com>