rfnoc: graph: Optimize property propagation algorithm
This introduces the concept of a resolution context, because the
property propagation algorithm needs to behave differently when called
during an initialization step (e.g. when the graph is committed), or
when the user changes a property on one of the nodes after it was
committed.

The algorithm is modified as follows:

- When called during an initialization step, then all nodes get resolved
  at least once. If nodes added new properties, then all nodes get
  touched again until the max number of iterations is reached.
- When called because a node modified one of its properties, then that
  node is always resolved first. From there, all other nodes are
  resolved in topological order. However, the algorithm immediately
  terminates as soon as there are no more dirty nodes.
- When called because a node modified one of its properties, but the
  graph is currently not in a committed state, then that node will do
  a local property resolution.