mpmd: Fix corner case in MPM device reachability
To check if a device is fully reachable through the
chdr address i.e. SFP port, an RPC request is sent
to "second_addr" and "addr" of that device. This commit
fixes the case where if "second_addr" is not connected to
the host, the function returns earlier and does not
check for "addr".